move UniverseObjectId and StarSystemCoordinates to their own files
