+++ src/ts/cosmosJourneyer.ts
@@ -36,7 +36,8 @@ import { Vector3 } from "@babylonjs/core/Maths/math.vector";
 import { Quaternion } from "@babylonjs/core/Maths/math";
 import { setRotationQuaternion } from "./uberCore/transforms/basicTransform";
 import { encodeBase64 } from "./utils/base64";
-import { StarSystemCoordinates, UniverseCoordinates } from "./utils/coordinates/universeCoordinates";
+import { UniverseCoordinates } from "./utils/coordinates/universeCoordinates";
+import { StarSystemCoordinates } from "./utils/coordinates/starSystemCoordinates";
 import { View } from "./utils/view";
 import { AudioManager } from "./audio/audioManager";
 import { AudioMasks } from "./audio/audioMasks";
