+++ src/ts/missions/common.ts
@@ -1,6 +1,6 @@
 import { Vector3 } from "@babylonjs/core/Maths/math.vector";
 import i18n from "../i18n";
-import { StarSystemCoordinates, starSystemCoordinatesEquals } from "../utils/coordinates/universeCoordinates";
+import { StarSystemCoordinates, starSystemCoordinatesEquals } from "../utils/coordinates/starSystemCoordinates";
 import { parseDistance } from "../utils/strings/parseToStrings";
 import { MissionContext } from "./missionContext";
 import { Settings } from "../settings";
