+++ src/ts/missions/nodes/actions/sightseeing/missionAsteroidFieldNodeSerialized.ts
@@ -15,7 +15,7 @@
 //  You should have received a copy of the GNU Affero General Public License
 //  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import { UniverseObjectIdSchema } from "../../../../utils/coordinates/universeCoordinates";
+import { UniverseObjectIdSchema } from "../../../../utils/coordinates/universeObjectId";
 import { z } from "zod";
 import { MissionNodeType } from "../../missionNodeType";
