+++ src/ts/missions/nodes/logic/missionOrNode.ts
@@ -20,7 +20,7 @@ import { MissionNodeType } from "../missionNodeType";
 import { MissionNodeBase } from "../missionNodeBase";
 import { MissionContext } from "../../missionContext";
 import i18n from "../../../i18n";
-import { StarSystemCoordinates } from "../../../utils/coordinates/universeCoordinates";
+import { StarSystemCoordinates } from "../../../utils/coordinates/starSystemCoordinates";
 import { StarSystemDatabase } from "../../../starSystem/starSystemDatabase";
 import type { MissionNode } from "../missionNode";
