+++ src/ts/missions/nodes/logic/missionXorNode.ts
@@ -19,7 +19,7 @@ import { MissionXorNodeSerialized } from "../missionNodeSerialized";
 import { MissionNodeType } from "../missionNodeType";
 import { MissionNodeBase } from "../missionNodeBase";
 import { MissionContext } from "../../missionContext";
-import { StarSystemCoordinates } from "../../../utils/coordinates/universeCoordinates";
+import { StarSystemCoordinates } from "../../../utils/coordinates/starSystemCoordinates";
 import { StarSystemDatabase } from "../../../starSystem/starSystemDatabase";
 import type { MissionNode } from "../missionNode";
