+++ src/ts/player/serializedPlayer.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 import { MissionSerializedSchema } from "../missions/missionSerialized";
 import { SpaceDiscoveryDataSchema } from "../society/encyclopaediaGalactica";
 import { SerializedSpaceshipSchema, DefaultSerializedSpaceship } from "../spaceship/serializedSpaceship";
-import { StarSystemCoordinatesSchema } from "../utils/coordinates/universeCoordinates";
+import { StarSystemCoordinatesSchema } from "../utils/coordinates/starSystemCoordinates";
 
 export const CompletedTutorialsSchema = z.object({
     stationLandingCompleted: z.boolean().default(false),
