+++ src/ts/ui/mainMenu.ts
@@ -28,7 +28,8 @@ import packageInfo from "../../../package.json";
 import i18n from "../i18n";
 import { Sounds } from "../assets/sounds";
 import { PanelType, SidePanels } from "./sidePanels";
-import { SystemObjectType, UniverseObjectId } from "../utils/coordinates/universeCoordinates";
+import { SystemObjectType } from "../utils/coordinates/universeCoordinates";
+import { UniverseObjectId } from "../utils/coordinates/universeObjectId";
 import { getObjectBySystemId } from "../utils/coordinates/orbitalObjectIdUtils";
 import { StarSystemModel, StarSystemModelUtils } from "../starSystem/starSystemModel";
 import { OrbitalObjectType } from "../architecture/orbitalObjectType";
