+++ src/ts/utils/getNeighborStarSystems.ts
@@ -16,7 +16,7 @@
 //  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import { Vector3 } from "@babylonjs/core/Maths/math.vector";
-import { StarSystemCoordinates, starSystemCoordinatesEquals } from "./coordinates/universeCoordinates";
+import { StarSystemCoordinates, starSystemCoordinatesEquals } from "./coordinates/starSystemCoordinates";
 import { Settings } from "../settings";
 import { StarSystemDatabase } from "../starSystem/starSystemDatabase";
