+++ tests/neighborStarSystems.spec.ts
@@ -17,7 +17,7 @@
 
 import { getNeighborStarSystemCoordinates } from "../src/ts/utils/getNeighborStarSystems";
 import { Vector3 } from "@babylonjs/core/Maths/math.vector";
-import { starSystemCoordinatesEquals } from "../src/ts/utils/coordinates/universeCoordinates";
+import { starSystemCoordinatesEquals } from "../src/ts/utils/coordinates/starSystemCoordinates";
 import { StarSystemDatabase } from "../src/ts/starSystem/starSystemDatabase";
 import { expect, test } from "vitest";
